enh: Flight plots moved to plots/flight_plots.py
