+++ rocketpy/plots/__init__.py
@@ -1,1 +1,1 @@
-from . import plots
+from .flight_plots import _FlightPlots
